+++ lib/moped/collection.rb
@@ -38,8 +38,8 @@ module Moped
         database.session.with(consistency: :strong) do |session|
           session.context.command(database.name, drop: name)
         end
-      rescue Moped::Errors::OperationFailure
-        false
+      rescue Moped::Errors::OperationFailure => e
+        raise e unless e.details["errmsg"] == "ns not found"
       end
     end