+++ spec/moped/collection_spec.rb
@@ -54,6 +54,27 @@ describe Moped::Collection do
         result = session[:users].drop
         result["ns"].should eq "moped_test.users"
       end
+
+      it "raises any error" do
+        Moped::Session::Context.any_instance.should_receive(:command).
+          with(session.send(:current_database).name, drop: "users").
+          and_raise("drop failed")
+        expect { session[:users].drop }.to raise_error "drop failed"
+      end
+
+      it "raises on Moped::Errors::OperationFailure" do
+        Moped::Session::Context.any_instance.should_receive(:command).
+          with(session.send(:current_database).name, drop: "users").
+          and_raise Moped::Errors::OperationFailure.new("drop", { "errmsg" => "unexpected"})
+        expect { session[:users].drop }.to raise_error Moped::Errors::OperationFailure, /unexpected/
+      end
+
+      it "doesn't raise on ns not found" do
+        Moped::Session::Context.any_instance.should_receive(:command).
+          with(session.send(:current_database).name, drop: "users").
+          and_raise Moped::Errors::OperationFailure.new("drop", { "errmsg" => "ns not found"})
+        expect { session[:users].drop }.to_not raise_error
+      end
     end
 
     context "when collection doesn't exist" do
@@ -65,6 +86,7 @@ describe Moped::Collection do
         session[:users].drop.should be_false
       end
     end
+
   end
 
   describe "#insert" do